api/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
firebase-admin
yfinance
pandas-ta
SQLAlchemy-Utils==0.41.1
celery[redis]
numpy<2.0

# Testing
pytest
pytest-cov
httpx
pytest-mock
pytest-asyncio
ruff